Tile Maps and Layers
I added in a very basic Tile set and added (I think correctly) collisions to the tiles. You may have to disable them or change the layers to have the real player interact with them in different ways because I didn't want to mess with the masks or layers until we got the player down. I added a collision box to the background for this reason, that I assumed that if we want to player to grab the background, we will need it to tell if he is touching the background. Of course, because I did not change the layer or mask, the player will run into like it is a solid wall. Feel free to change that if you know which needs to be changed. I also theoretically made the "ledges" one way, meaning you should be able to enter them. I may have also made the "tile_layers.tscn" the main scene on accident while running a test. If it is till the main scene, feel free to undo that.