Added delay for stamina regen.
Delay was added for stamina regen, once again using a timer and a bool. I doubt it is the best way to do that. However, here we are. Also, in retrospect my naming of the "stamina_drain" var is a little counterintuitive as it is also the stamina regen rate...feel free to change the name of the var.