Added rough timer and rough game over.
Added a timer node with text as well as a very basic game over screen with a button to restart for when time runs out. Used UI nodes for all of them as of right now. Also added debug buttons to stop and start the timer. To stop press ` (tilda) and to start after you have stopped it press \ (backslash).