Added Stamina Supplement item.
Added an item that stops the stamina drain. I think we could change this to be an item that increases stamina regen instead if we wanted to. Or maybe slows stamina drain. Also added a stamina_regen var to the player for ease of testing and messing with for items. Maybe will want to cut down once we have a better idea of things? Still.