Basic next level functionality added.
Added a way to move the scene to the next level while taking the player through the item selection screen. There is a var in global_data that tracks the level and when the player crosses the finish line we increment the var by 1. Then, in the item selection script, we get what the var is equal to with a match statement and assign a value to a var based on that number in the global data. This will probably have bugs